requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
jinja2==3.1.2
python-multipart==0.0.6
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2 